consolidate server entry
