--- server/minimal-server.js
@@ -1,44 +1,0 @@
-import express from 'express';
-import dotenv from 'dotenv';
-
-// Load environment variables
-dotenv.config();
-
-const app = express();
-const PORT = process.env.PORT || 5000;
-
-// Basic middleware
-app.use(express.json());
-
-// Simple health check
-app.get('/health', (req, res) => {
-  res.json({ 
-    status: 'ok', 
-    message: 'NestMap Server is running',
-    timestamp: new Date().toISOString(),
-    environment: process.env.NODE_ENV || 'development'
-  });
-});
-
-// Simple test route
-app.get('/test', (req, res) => {
-  res.json({ message: 'Test endpoint working' });
-});
-
-// Start server
-app.listen(PORT, 'localhost', () => {
-  console.log(`🚀 Minimal server running on http://localhost:${PORT}`);
-  console.log(`📋 Health check: http://localhost:${PORT}/health`);
-  console.log(`🧪 Test endpoint: http://localhost:${PORT}/test`);
-});
-
-// Handle errors
-process.on('uncaughtException', (error) => {
-  console.error('Uncaught Exception:', error);
-  process.exit(1);
-});
-
-process.on('unhandledRejection', (reason, promise) => {
-  console.error('Unhandled Rejection at:', promise, 'reason:', reason);
-  process.exit(1);
-});
