+++ server/src/main.ts
@@ -1,0 +1,133 @@
+import 'dotenv/config';
+import express from 'express';
+import cors from 'cors';
+import helmet from 'helmet';
+import { logger } from '../utils/logger';
+
+// Import authentication middleware
+import { authenticate } from '../middleware/secureAuth';
+
+// Import route handlers
+import authRoutes from '../routes/auth-simple';
+import flightRoutes from '../routes/flights';
+import organizationRoutes from '../routes/organizations';
+import tripRoutes from '../routes/trips';
+
+// Create Express app
+const app = express();
+
+// Configuration
+const PORT = process.env.PORT || 3000;
+const NODE_ENV = process.env.NODE_ENV || 'development';
+const ALLOWED_ORIGINS = process.env.ALLOWED_ORIGINS?.split(',') || ['http://localhost:3000', 'http://localhost:5173'];
+
+// Basic security middleware
+app.use(helmet({
+  contentSecurityPolicy: {
+    directives: {
+      defaultSrc: ["'self'"],
+      styleSrc: ["'self'", "'unsafe-inline'"],
+      scriptSrc: ["'self'"],
+      imgSrc: ["'self'", "data:", "https:"],
+      connectSrc: ["'self'", "https://api.duffel.com"],
+      fontSrc: ["'self'"],
+      objectSrc: ["'none'"],
+      mediaSrc: ["'self'"],
+      frameSrc: ["'none'"],
+    },
+  },
+  crossOriginEmbedderPolicy: false,
+}));
+
+// CORS configuration
+app.use(cors({
+  origin: ALLOWED_ORIGINS,
+  credentials: true,
+  methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+  allowedHeaders: ['Content-Type', 'Authorization', 'X-Requested-With'],
+}));
+
+// Body parsing middleware
+app.use(express.json({ limit: '10mb' }));
+app.use(express.urlencoded({ extended: true, limit: '10mb' }));
+
+// Request logging
+if (NODE_ENV === 'development') {
+  app.use((req, res, next) => {
+    logger.info(`${req.method} ${req.path}`, {
+      ip: req.ip,
+      userAgent: req.headers['user-agent']
+    });
+    next();
+  });
+}
+
+// Public routes (no authentication required)
+app.use('/api/auth', authRoutes);
+
+// Protected routes (require JWT authentication)
+app.use('/api/flights', authenticate, flightRoutes);
+app.use('/api/organizations', authenticate, organizationRoutes);
+app.use('/api/trips', authenticate, tripRoutes);
+
+// Health check route
+app.get('/health', (_req, res) => {
+  res.json({
+    status: 'healthy',
+    timestamp: new Date().toISOString(),
+    version: '1.0.0',
+    environment: NODE_ENV
+  });
+});
+
+// Global error handler
+app.use((error: any, req: express.Request, res: express.Response, next: express.NextFunction) => {
+  logger.error('Unhandled error:', error);
+  
+  if (NODE_ENV === 'production') {
+    res.status(500).json({
+      error: 'Internal Server Error',
+      message: 'An unexpected error occurred'
+    });
+  } else {
+    res.status(500).json({
+      error: 'Internal Server Error',
+      message: error.message,
+      stack: error.stack
+    });
+  }
+});
+
+// 404 handler
+app.use('*', (req, res) => {
+  res.status(404).json({
+    error: 'Not Found',
+    message: `Route ${req.originalUrl} not found`
+  });
+});
+
+// Start server
+const server = app.listen(PORT, () => {
+  logger.info(`🚀 NestMap API Server running on port ${PORT}`);
+  logger.info(`📚 Environment: ${NODE_ENV}`);
+  logger.info(`🔗 Health check: http://localhost:${PORT}/health`);
+});
+
+// Graceful shutdown
+process.on('SIGTERM', () => {
+  logger.info('SIGTERM received, shutting down gracefully');
+  server.close(() => {
+    logger.info('Server closed');
+    process.exit(0);
+  });
+});
+
+process.on('SIGINT', () => {
+  logger.info('SIGINT received, shutting down gracefully');
+  server.close(() => {
+    logger.info('Server closed');
+    process.exit(0);
+  });
+});
+
+export default app;
